completed benchmark - initial
